+++ src/main/java/PKUCRProject/PKUCR/backend/Dao/UserMapper.java
@@ -14,7 +14,7 @@ public interface UserMapper {
 
     @Insert("insert into users (email, password, permission) values (#{email}, #{password}, #{permission})")
     @Options(useGeneratedKeys = true, keyProperty = "id", keyColumn = "id")
-    int insert(User user);
+    Long insert(User user);
 
     @Select("select * from users where email = #{email}")
     User selectByEmail(String email);
@@ -23,12 +23,12 @@ public interface UserMapper {
     User selectByEmailAndPassword(String email, String password);
 
     @Select("select * from users where id = #{id}")
-    User selectById(int id);
+    User selectById(Long id);
 
     @Update("update users set email = #{email}, password = #{password}, permission = #{permission} where id = #{id}")
     void update(User user);
 
     @Delete("delete from users where id = #{id}")
-    void delete(int id);
+    void delete(Long id);
     
 }
